test(frontend): isolate BarGauge hosted frame checks
